# Improve app summary performance