--- a/corehq/apps/app_manager/templates/app_manager/form_summary.html
+++ b/corehq/apps/app_manager/templates/app_manager/form_summary.html
@@ -54,7 +54,7 @@
     <!-- /ko -->
 
     <ul class="fa-ul" data-bind="foreach: modules">
-        <li data-bind="slideVisible: isVisible">
+        <li data-bind="visible: isVisible">
             <h4>
                 <a data-bind="ifnot: $root.readOnly, attr: { href: url }">
                     <i data-bind="attr: { 'class': icon }"></i>
@@ -69,7 +69,7 @@
             <!-- ko template: {name: 'question-attribute', data: { attribute: module_filter, icon: 'fa-code-fork',
                  visibleObs: $root.showRelevance } }--><!-- /ko -->
             <ul class="fa-ul" data-bind="foreach: forms">
-                <li data-bind="slideVisible: isVisible">
+                <li data-bind="visible: isVisible">
                     <h5>
                         <a data-bind="ifnot: $root.readOnly, attr: { href: url }">
                             <i data-bind="attr: { 'class': icon }"></i>
@@ -84,7 +84,7 @@
                     <!-- ko template: {name: 'question-attribute', data: { attribute: form_filter, icon: 'fa-code-fork',
                          visibleObs: $root.showRelevance } }--><!-- /ko -->
                     <ol data-bind="foreach: questions">
-                        <li data-bind="slideVisible: isVisible">
+                        <li data-bind="visible: isVisible">
                             <i data-bind="attr: { 'class': $root.questionIcon($data) }" title="type"></i>
                             <span data-bind="visible: $root.showIds, text: value"></span>
                             <span data-bind="visible: $root.showLabels, text: $root.translateQuestion($data)"></span>
@@ -113,7 +113,7 @@
 
     {# Button to toggle question attribute #}
     <script type="text/html" id="question-attribute">
-        <ul class="fa-ul" data-bind="if: attribute, slideVisible: visibleObs">
+        <ul class="fa-ul" data-bind="if: attribute, visible: visibleObs">
             <li>
                 <i data-bind="attr: {'class': 'fa-li fa text-muted ' + icon}"></i>
                 <span data-bind="text: attribute"></span>