Very large apps sometimes crash the browser when viewed in the form summary (and if they do load, they take an extremely long time, and peg the CPU).

- Most importantly, this removes slideVisible which made the eCHIS app able to load (before it would just crash my browser, and sometimes my computer, and peg the CPU at 100%)
- Removes the use of templates in loops (I got the idea from here: https://stackoverflow.com/a/9723763/2957657), this reduced the time to load call applyBindings for the eCHIS app from 24s to 15s
- It also removes nodes from the DOM which will never have data in them, as the number of nodes was becoming insane when there are thousands of questions. This improved scroll performance and the responsiveness of the page when clicking the sidebar links.
- Removes extraneous span nodes which also add thousands of DOM nodes

For the eCHIS app, this now takes 15-20 seconds to load, which is still not good, but better late than never.

There is no significant difference for very small apps that I could see (reduces page load time from 236ms to 114ms, but this isn't noticeable)