@proteusvacuum 2019-01-21
Still need to do a pass on the case summary, but it also now loads for eCHIS (probably thanks to the change slideVisible)