@orangejenny 2019-01-21
Disappointing that the animations are such a burden, since they make this page less jumpy, but so it goes.